# Conflicts:
#	cmd/manager/validation.go
#	pkg/controller/logstash/pod_test.go

+++ pkg/controller/logstash/network/ports.go
@@ -6,5 +6,5 @@ package network
 
 const (
 	// HTTPPort is the (default) API port used by Logstash
-	HTTPPort = 9600
+	HTTPPort = int32(9600)
 )
